Migrate reading suggestions store to Pinia
